Added uniform buffer impl to cpp file
